moved everything to back Main.cpp
