+++ main.cpp
@@ -153,7 +153,124 @@ object
 */
 #include <iostream>
 #include <cmath>
-#include "notMain.h"
+
+
+struct FloatType;
+struct DoubleType;
+struct IntType;
+
+
+struct FloatType // wrote UDT part 1 - 1)
+{
+    FloatType(float _float); //Part 1 - 4) 
+    ~FloatType();
+    operator float() { return *ownedVariable; }
+    FloatType& pow(const IntType&); 
+    FloatType& pow(const FloatType&); 
+    FloatType& pow(const DoubleType&); 
+    FloatType& pow(float); 
+    FloatType& powInternal(); 
+private: // Part 3 - 3)
+    float* ownedVariable;
+};
+
+struct DoubleType
+{
+    DoubleType(double _double);
+    ~DoubleType();
+    operator double() { return *ownedVariable; }
+    DoubleType& pow(const IntType&);
+    DoubleType& pow(const FloatType&);
+    DoubleType& pow(const DoubleType&);
+    DoubleType& pow(double);
+    DoubleType& powInternal();
+private:
+    double* ownedVariable;
+};
+
+struct IntType
+{
+    IntType(int _int);
+    ~IntType();
+    operator int() { return *ownedVariable; }
+    IntType& pow(const IntType&);
+    IntType& pow(const FloatType&);
+    IntType& pow(const DoubleType&);
+    IntType& pow(int);
+    IntType& powInternal();
+private:
+    int* ownedVariable;
+};
+
+
+struct Point
+{
+Point(float, float) : x{0}, y {0}; //Part 6 - 7)
+Point(FloatType);
+    Point& multiply(float m)
+    {
+        x *= m;
+        y *= m;
+        return *this;
+    }
+    void toString()
+    {
+        std::cout << "The x coordinate is: " << x << "\nThe y coordinate is: " << y << std::endl;
+    }
+private:
+    float x{0}, y{0};
+};
+
+
+
+//moved implemetations out of class. part 3 - 1)
+FloatType::FloatType(float _float) : ownedVariable(new float(_float)) { } // made heap-allocated primitive type. Part 1 -2) 
+FloatType::~FloatType() { delete ownedVariable; ownedVariable = nullptr; }// prevents heap allocating type from leaking Part1 - 3)
+//void FloatType::float() {return *ownedVariable} // - tried to move this function out of class, not sure why I can't
+
+
+DoubleType::DoubleType(double _double) : ownedVariable(new double(_double)) { }
+DoubleType::~DoubleType() {delete ownedVariable; ownedVariable = nullptr; }
+//void DoubleType::double() {return *ownedVariable}
+
+
+IntType::IntType(int _int) : ownedVariable(new int (_int)){ }
+IntType::~IntType() { delete ownedVariable; ownedVariable = nullptr;}
+//void IntType::int() {return *ownedVariable}
+
+Point::Point();
+
+//Point::Point(FloatType) {}
+//Point::Point(DoubleType) {}
+//Point::Point(IntType) {}
+//         Free Functions               
+
+void updateValue( float& value)
+{
+    value += value;
+}
+
+void updateValue( double& value)
+{
+    value += value;
+}
+
+void updateValue( int& value)
+{
+    value += value;
+}
+
+void divider()
+{
+    std::cout << "\n\n===============================\n\n";
+}
+
+
+
+
+
+
+
